Fixed issue with socketio and bootstrap-angular.
Templates for bootstrap angular are based on ng-template tag instead of template now. Changed as appropriate.

Socketio is no longer compatible with newer versions of angular as they have breaking changes in the use of the "global"
value in the window. Shim sets a global variable to the entire window.